+++ sdk/cosmos/azure-cosmos/src/main/java/com/azure/cosmos/models/ModelBridgeInternal.java
@@ -435,6 +435,8 @@ public final class ModelBridgeInternal {
             ((PartitionKeyDefinition) t).populatePropertyBag();
         } else if (t instanceof SpatialSpec) {
             ((SpatialSpec) t).populatePropertyBag();
+        } else if (t instanceof CosmosVectorIndexSpec) {
+            ((CosmosVectorIndexSpec) t).populatePropertyBag();
         } else if (t instanceof SqlParameter) {
             ((SqlParameter) t).populatePropertyBag();
         } else if (t instanceof SqlQuerySpec) {
@@ -468,6 +470,8 @@ public final class ModelBridgeInternal {
             return ((PartitionKeyDefinition) t).getJsonSerializable();
         } else if (t instanceof SpatialSpec) {
             return ((SpatialSpec) t).getJsonSerializable();
+        } else if (t instanceof CosmosVectorIndexSpec) {
+            return ((CosmosVectorIndexSpec) t).getJsonSerializable();
         } else if (t instanceof SqlParameter) {
             return ((SqlParameter) t).getJsonSerializable();
         } else if (t instanceof SqlQuerySpec) {